+++ features/admin.core.v1/models/config.ts
@@ -458,10 +458,6 @@ export interface UIConfigInterface extends CommonUIConfigInterface<FeatureConfig
      * Configurations for IDP templates.
      */
     connectionTemplates?: any;
-    /**
-     * Config if beta tag should be displayed for sms otp for password recovery feature.
-     */
-    showSmsOtpPwdRecoveryFeatureStatusChip?: boolean;
 }
 
 /**
