+++ features/admin.server-configurations.v1/constants/governance-connector-constants.ts
@@ -60,40 +60,12 @@ export class GovernanceConnectorConstants {
         EXPIRY_TIME_MAX_VALUE: number;
         EXPIRY_TIME_MIN_LENGTH: number;
         EXPIRY_TIME_MIN_VALUE: number;
-        MAX_FAILED_ATTEMPT_COUNT_MIN_LENGTH: number;
-        MAX_FAILED_ATTEMPT_COUNT_MAX_LENGTH: number;
-        MAX_FAILED_ATTEMPT_COUNT_MIN_VALUE: number;
-        MAX_FAILED_ATTEMPT_COUNT_MAX_VALUE: number;
-        MAX_RESEND_COUNT_MIN_LENGTH: number;
-        MAX_RESEND_COUNT_MAX_LENGTH: number;
-        MAX_RESEND_COUNT_MIN_VALUE: number;
-        MAX_RESEND_COUNT_MAX_VALUE: number;
-        SMS_OTP_EXPIRY_TIME_MAX_LENGTH: number;
-        SMS_OTP_EXPIRY_TIME_MAX_VALUE: number;
-        SMS_OTP_CODE_LENGTH_MAX_LENGTH: number;
-        SMS_OTP_CODE_LENGTH_MAX_VALUE: number;
-        SMS_OTP_CODE_LENGTH_MIN_LENGTH: number;
-        SMS_OTP_CODE_LENGTH_MIN_VALUE: number;
     } = {
 
         EXPIRY_TIME_MAX_LENGTH: 5,
         EXPIRY_TIME_MAX_VALUE: 10080,
         EXPIRY_TIME_MIN_LENGTH: 1,
-        EXPIRY_TIME_MIN_VALUE: 1,
-        MAX_FAILED_ATTEMPT_COUNT_MAX_LENGTH: 2,
-        MAX_FAILED_ATTEMPT_COUNT_MAX_VALUE: 10,
-        MAX_FAILED_ATTEMPT_COUNT_MIN_LENGTH: 1,
-        MAX_FAILED_ATTEMPT_COUNT_MIN_VALUE: 1,
-        MAX_RESEND_COUNT_MAX_LENGTH: 1,
-        MAX_RESEND_COUNT_MAX_VALUE: 5,
-        MAX_RESEND_COUNT_MIN_LENGTH: 1,
-        MAX_RESEND_COUNT_MIN_VALUE: 1,
-        SMS_OTP_CODE_LENGTH_MAX_LENGTH: 2,
-        SMS_OTP_CODE_LENGTH_MAX_VALUE: 10,
-        SMS_OTP_CODE_LENGTH_MIN_LENGTH: 1,
-        SMS_OTP_CODE_LENGTH_MIN_VALUE: 6,
-        SMS_OTP_EXPIRY_TIME_MAX_LENGTH: 4,
-        SMS_OTP_EXPIRY_TIME_MAX_VALUE: 1440
+        EXPIRY_TIME_MIN_VALUE: 1
     };
 
     /**
