+++ features/admin.server-configurations.v1/constants/server-configurations-constants.ts
@@ -238,20 +238,6 @@ export class ServerConfigurationsConstants {
 	public static readonly PASSWORD_RECOVERY_QUESTION_FORCED_ENABLE: string =
 		"Recovery.Question.Password.Forced.Enable";
 
-	public static readonly RECOVERY_EMAIL_LINK_ENABLE: string = "Recovery.Notification.Password.emailLink.Enable";
-	public static readonly RECOVERY_SMS_OTP_ENABLE: string = "Recovery.Notification.Password.smsOtp.Enable";
-	public static readonly RECOVERY_OTP_USE_UPPERCASE: string =
-		"Recovery.Notification.Password.OTP.UseUppercaseCharactersInOTP";
-
-	public static readonly RECOVERY_OTP_USE_LOWERCASE: string =
-		"Recovery.Notification.Password.OTP.UseLowercaseCharactersInOTP";
-
-	public static readonly RECOVERY_OTP_USE_NUMERIC: string = "Recovery.Notification.Password.OTP.UseNumbersInOTP";
-	public static readonly RECOVERY_OTP_LENGTH: string = "Recovery.Notification.Password.OTP.OTPLength";
-	public static readonly RECOVERY_MAX_RESEND_COUNT: string = "Recovery.Notification.Password.MaxResendAttempts";
-	public static readonly RECOVERY_MAX_FAILED_ATTEMPTS_COUNT: string =
-		"Recovery.Notification.Password.MaxFailedAttempts";
-
 	/**
 	 * Login policies - account locking API Keyword constants.
 	 */
