+++ features/admin.server-configurations.v1/forms/connector-form-factory.tsx
@@ -26,7 +26,7 @@ import { AnalyticsConfigurationForm } from "./analytics-form";
 import { AskPasswordForm } from "./ask-password";
 import { LoginAttemptSecurityConfigurationFrom } from "./login-attempt-security-form";
 import { MultiAttributeLoginForm } from "./multi-attribute-login";
-import { PasswordRecoveryConfigurationForm } from "./password-recovery-form/password-recovery-form";
+import { PasswordRecoveryConfigurationForm } from "./password-recovery-form";
 import { SelfRegistrationForm } from "./self-registration-form";
 import { UsernameRecoveryConfigurationForm } from "./username-recovery-form";
 import { AppState, FeatureConfigInterface, history } from "../../admin.core.v1";
