--- features/admin.server-configurations.v1/forms/password-recovery-form/password-recovery-form.scss
@@ -1,29 +1,0 @@
-$form-level-2-margin-left: 50px;
-
-.connector-form.password-recovery-form {
-    h4.ui.header.heading {
-        margin-bottom: 35px;
-    }
-
-    h5.ui.header.heading {
-        margin-bottom: 15px;
-    }
-
-    .ui-hint {
-        margin-bottom: 25px;
-    }
-
-    form > div > div:nth-child(5),
-        div:nth-child(6),
-        div:nth-child(7),
-        div:nth-child(8),
-        div:nth-child(12),
-        div:nth-child(13) {
-        margin-left: 50px;
-    }
-
-    .MuiChip-root {
-        margin-left: 10px !important;
-    }
-
-}
