+++ features/admin.server-configurations.v1/pages/connector-edit-page.tsx
@@ -287,10 +287,7 @@ export const ConnectorEditPage: FunctionComponent<ConnectorEditPageInterface> = 
 
         if (
             serverConfigurationConfig.connectorToggleName[ connector?.name ] &&
-            serverConfigurationConfig.autoEnableConnectorToggleProperty &&
-            // Recovery connector does not use  a connector property. If either email link or
-            // sms-otp recovery is enabled, the connector is considered to be enabled.
-            connectorId !== ServerConfigurationsConstants.ACCOUNT_RECOVERY_CONNECTOR_ID
+            serverConfigurationConfig.autoEnableConnectorToggleProperty
         ) {
             data.properties.push({
                 name: GovernanceConnectorUtils.decodeConnectorPropertyName(
@@ -575,6 +572,10 @@ export const ConnectorEditPage: FunctionComponent<ConnectorEditPageInterface> = 
                 return ServerConfigurationsConstants.SELF_REGISTRATION_ENABLE;
             case ServerConfigurationsConstants.CAPTCHA_FOR_SSO_LOGIN_CONNECTOR_ID:
                 return ServerConfigurationsConstants.RE_CAPTCHA_AFTER_MAX_FAILED_ATTEMPTS_ENABLE;
+            case ServerConfigurationsConstants.ACCOUNT_RECOVERY_CONNECTOR_ID:
+                return type === "username"
+                    ? undefined
+                    : ServerConfigurationsConstants.PASSWORD_RECOVERY_NOTIFICATION_BASED_ENABLE;
             case ServerConfigurationsConstants.ORGANIZATION_SELF_SERVICE_CONNECTOR_ID:
                 return ServerConfigurationsConstants.ORGANIZATION_SELF_SERVICE_ENABLE;
             case ServerConfigurationsConstants.MULTI_ATTRIBUTE_LOGIN_CONNECTOR_ID:
